refactor(assertions): simplify assertion calls by removing unnecessary casts

+++ TUnit.Assertions.Tests/TypeOfTests.cs
@@ -242,7 +242,7 @@ public class TypeOfTests
 
         // Should be able to use collection assertion methods
         await Assert.That(nullableBytes).HasCount(5);
-        await Assert.That(nullableBytes).Contains((byte)3);
+        await Assert.That(nullableBytes).Contains(3);
         await Assert.That(nullableBytes).IsInOrder();
 
     }
